backend/requirements.txt
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
django-q2==1.6.2
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
sqlparse==0.5.4
tzdata==2025.2
virtualenv==20.35.3
